remove kvclient v2 related code .
